# modelindexer: Reduce locking on flushActive