@stuartnelson3 2022-03-28
confirmed, results are below:

# 8.1.2
with patch b951097af2cd622b07b48e1cd77dbd84c1addf54
```
# first run
BenchmarkAgentGo-4    	      78	 916482566 ns/op	         0 error_responses/sec	       114.6 errors/sec	      5713 events/sec	       264.2 metrics/sec	      3768 spans/sec	      1567 txs/sec	 1043712 B/op	    1958 allocs/op
BenchmarkAgentNodeJS-4	     134	 507775279 ns/op	         0 error_responses/sec	        96.50 errors/sec	      3965 events/sec	       237.1 metrics/sec	      2090 spans/sec	      1542 txs/sec	 1903909 B/op	    2949 allocs/op
BenchmarkAgentPython-4	      44	1551750409 ns/op	         0 error_responses/sec	        46.40 errors/sec	      4528 events/sec	      1369 metrics/sec	      2629 spans/sec	       484.6 txs/sec	 5314182 B/op	    8721 allocs/op
BenchmarkAgentRuby-4  	      81	 766699588 ns/op	         0 error_responses/sec	       174.8 errors/sec	      4898 events/sec	       361.2 metrics/sec	      2880 spans/sec	      1482 txs/sec	 1515841 B/op	    2530 allocs/op
# second run
BenchmarkAgentGo-4    	      66	1033106531 ns/op	         0 error_responses/sec	       101.6 errors/sec	      5076 events/sec	       242.1 metrics/sec	      3342 spans/sec	      1390 txs/sec	 1063627 B/op	    1970 allocs/op
BenchmarkAgentNodeJS-4	     122	 501214184 ns/op	         0 error_responses/sec	        97.76 errors/sec	      4017 events/sec	       240.5 metrics/sec	      2117 spans/sec	      1562 txs/sec	 1910840 B/op	    2968 allocs/op
BenchmarkAgentPython-4	      46	1488563862 ns/op	         0 error_responses/sec	        48.37 errors/sec	      4720 events/sec	      1427 metrics/sec	      2740 spans/sec	       505.2 txs/sec	 5291884 B/op	    8568 allocs/op
BenchmarkAgentRuby-4  	      94	 801948566 ns/op	         0 error_responses/sec	       167.1 errors/sec	      4682 events/sec	       345.0 metrics/sec	      2753 spans/sec	      1417 txs/sec	 1497904 B/op	    2453 allocs/op
# third run
BenchmarkAgentGo-4    	      67	 996238061 ns/op	         0 error_responses/sec	       105.4 errors/sec	      5256 events/sec	       243.4 metrics/sec	      3466 spans/sec	      1441 txs/sec	 1060362 B/op	    2007 allocs/op
BenchmarkAgentNodeJS-4	     123	 530209145 ns/op	         0 error_responses/sec	        92.42 errors/sec	      3800 events/sec	       229.9 metrics/sec	      2001 spans/sec	      1477 txs/sec	 1910951 B/op	    2979 allocs/op
BenchmarkAgentPython-4	      42	1580797452 ns/op	         0 error_responses/sec	        45.55 errors/sec	      4443 events/sec	      1341 metrics/sec	      2580 spans/sec	       475.7 txs/sec	 5313029 B/op	    8645 allocs/op
BenchmarkAgentRuby-4  	      91	 802732008 ns/op	         0 error_responses/sec	       166.9 errors/sec	      4677 events/sec	       344.7 metrics/sec	      2751 spans/sec	      1415 txs/sec	 1502372 B/op	    2473 allocs/op
```

# 8.1.1
02b37b77dd34ddb2b13d99bb0f87f07b04e22d32
```
# first run
BenchmarkAgentGo-4    	      21	3345064759 ns/op	         0 error_responses/sec	        31.39 errors/sec	      1567 events/sec	        74.12 metrics/sec	      1032 spans/sec	       429.3 txs/sec	 1311027 B/op	    2887 allocs/op
BenchmarkAgentNodeJS-4	      39	1723084867 ns/op	         0 error_responses/sec	        28.44 errors/sec	      1170 events/sec	        71.84 metrics/sec	       615.8 spans/sec	       454.4 txs/sec	 2036449 B/op	    3177 allocs/op
BenchmarkAgentPython-4	      13	5067959232 ns/op	         0 error_responses/sec	        14.21 errors/sec	      1387 events/sec	       419.9 metrics/sec	       804.9 spans/sec	       148.4 txs/sec	 5849044 B/op	   10527 allocs/op
BenchmarkAgentRuby-4  	      24	2661627765 ns/op	         0 error_responses/sec	        50.35 errors/sec	      1412 events/sec	       105.8 metrics/sec	       829.6 spans/sec	       426.8 txs/sec	 1762924 B/op	    3322 allocs/op
# second run
BenchmarkAgentGo-4    	      20	3409856369 ns/op	         0 error_responses/sec	        30.79 errors/sec	      1537 events/sec	        72.83 metrics/sec	      1013 spans/sec	       421.1 txs/sec	 1421068 B/op	    3870 allocs/op
BenchmarkAgentNodeJS-4	      38	1759422297 ns/op	         0 error_responses/sec	        27.85 errors/sec	      1146 events/sec	        70.43 metrics/sec	       603.0 spans/sec	       445.0 txs/sec	 2065048 B/op	    3523 allocs/op
BenchmarkAgentPython-4	      12	5200090475 ns/op	         0 error_responses/sec	        13.85 errors/sec	      1358 events/sec	       414.9 metrics/sec	       784.4 spans/sec	       144.6 txs/sec	 5923609 B/op	   10946 allocs/op
BenchmarkAgentRuby-4  	      25	2745435045 ns/op	         0 error_responses/sec	        48.81 errors/sec	      1372 events/sec	       104.7 metrics/sec	       804.2 spans/sec	       413.8 txs/sec	 1762541 B/op	    3410 allocs/op
# third run
BenchmarkAgentGo-4    	      22	3154050917 ns/op	         0 error_responses/sec	        33.29 errors/sec	      1662 events/sec	        78.50 metrics/sec	      1095 spans/sec	       455.3 txs/sec	 1295200 B/op	    2845 allocs/op
BenchmarkAgentNodeJS-4	      48	1565858554 ns/op	         0 error_responses/sec	        31.29 errors/sec	      1287 events/sec	        78.48 metrics/sec	       677.6 spans/sec	       500.0 txs/sec	 2018854 B/op	    3273 allocs/op
BenchmarkAgentPython-4	      15	4442788288 ns/op	         0 error_responses/sec	        16.21 errors/sec	      1582 events/sec	       478.6 metrics/sec	       918.1 spans/sec	       169.3 txs/sec	 5826198 B/op	   10860 allocs/op
BenchmarkAgentRuby-4  	      24	2546478761 ns/op	         0 error_responses/sec	        52.62 errors/sec	      1479 events/sec	       113.1 metrics/sec	       867.1 spans/sec	       446.1 txs/sec	 1779016 B/op	    3480 allocs/op
```